feat(project_iam_member.tf): add Terraform configuration for adding a Google IAM member to a project
This commit adds a new Terraform configuration file, `project_iam_member.tf`, which defines a resource block for adding a Google IAM member to a project. The IAM member is specified as a user with the email address `google@fajarmaftuhfadli.com`. The member is assigned the `roles/storage.admin` role.

This configuration allows for easy management of IAM members in the project, providing the necessary permissions for the specified user.